backend/requirements.txt
yfinance>=0.2.3
pandas>=2.0.0
numpy>=1.24.0
Flask>=3.0.0
flask-cors>=4.0.0
scipy>=1.11.0
stockstats>=0.6.0
pytest>=7.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
schedule>=1.2.0
PyYAML>=6.0.0
pyarrow>=14.0.0
